Allow linking to versioned libperl
When trying to link to the perlapi, the module gets the library name
from $Config{libperl}, and then performs some transformations to
convert the value into a linker flag.

This transformation expects an unversioned library name (libperl.so) during this
transformation:
Cut off the first "lib", and remove everything after the last dot: libperl.so -> perl

This transformation doesn't work in case the library is versioned, e.g libperl.so.5.40.2.
In this case the linker flag becomes "perl.so.5.40", which is not a valid library name,
and the linking fails.

To avoid it, this patch changes the second step of the transformation logic:
instead of cutting off everything after the last dot, cut off everything after the first one.

Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>